# perf(bundler): fine-tuning of Webpack/Rspack optimizations