--- a/packages/docusaurus-bundler/src/index.ts
+++ b/packages/docusaurus-bundler/src/index.ts
@@ -12,6 +12,7 @@ export {
   getCSSExtractPlugin,
   getCopyPlugin,
   getProgressBarPlugin,
+  registerBundlerTracing,
 } from './currentBundler';
 
 export {getMinimizers} from './minification';